refactor: PointGenerator 테스트를 위한 정적펙토리메서드 제거
