+++ src/main/java/Application.java
@@ -1,8 +1,11 @@
 import controller.LadderGameController;
+import domain.util.PointGenerator;
+import domain.util.RandomPointGenerator;
 
 public class Application {
 	public static void main(String[] args) {
-		LadderGameController ladderGameController = new LadderGameController();
+		PointGenerator pointGenerator = new RandomPointGenerator();
+		LadderGameController ladderGameController = new LadderGameController(pointGenerator);
 		ladderGameController.runGame();
 	}
 }
