+++ src/main/java/domain/util/PointGenerator.java
@@ -1,12 +1,6 @@
 package domain.util;
 
 public interface PointGenerator {
-	static PointGenerator getInstance(boolean isRandom) {
-		if (isRandom) {
-			return new RandomPointGenerator();
-		}
-		return new FixedPresencePointGenerator();
-	}
 
 	Point generate();
 }
