+++ src/test/java/domain/FixedPresencePointGenerator.java
@@ -1,4 +1,7 @@
-package domain.util;
+package domain;
+
+import domain.util.Point;
+import domain.util.PointGenerator;
 
 public class FixedPresencePointGenerator implements PointGenerator {
     @Override
